選択肢表示エリアをWidget化し、ゲーム画面に組み込むためのコード修正
