+++ lib/game_screen.dart
@@ -4,6 +4,7 @@ import 'package:flutter_nobel_app/usecase/admob_usecase.dart';
 import 'package:flutter_nobel_app/usecase/choice_usecase.dart';
 import 'package:flutter_nobel_app/usecase/save_usecase.dart';
 import 'package:flutter_nobel_app/usecase/story_usecase.dart';
+import 'package:flutter_nobel_app/widget/choose_screen_widget.dart';
 import 'package:flutter_nobel_app/widget/image_screen_widget.dart';
 import 'package:flutter_nobel_app/widget/speaker_area_widget.dart';
 import 'package:flutter_nobel_app/widget/text_area_widget.dart';
@@ -72,35 +73,9 @@ class _GameScreenState extends State<GameScreen> {
                     SpeakerAreaWidget(allStory: widget.allStory, storyUsecase: storyUsecase),
                                     
                   // 選択肢表示エリア
-                  AnimatedOpacity(
-                    opacity: usecase.isChoice ? 1.0 : 0.0,
-                    duration: const Duration(milliseconds: 1000),
-                    child: IgnorePointer(
-                      ignoring: !usecase.isChoice,
-                      child: Container(
-                        color: Colors.black.withAlpha(180),
-                        width: double.infinity,
-                        height: double.infinity,
-                        child: Center(
-                          child: Column(
-                            mainAxisAlignment: MainAxisAlignment.center,
-                            children: [
-                              ...usecase.currentChoice.map((choice) {
-                                return Padding(
-                                  padding: const EdgeInsets.all(8.0),
-                                  child: ElevatedButton(
-                                    onPressed: () {
-                                      usecase.tabSelect(choice, widget.allStory);
-                                    },
-                                    child: Text(choice.word)
-                                  ),
-                                );
-                              })
-                            ]
-                          ),
-                        ),
-                      ),
-                    ),
+                  ChooseScreenWidget(
+                    usecase: usecase,
+                    allStory: widget.allStory
                   )
                 ],
               ),
